requirements.txt
fastapi
uvicorn[standard]
requests
python-dotenv
psycopg2-binary
sqlalchemy
yfinance
pandas
numpy
beautifulsoup4
schedule
matplotlib
